Fix on lsh distances

+++ lib/Helper_Functions.h
@@ -12,6 +12,6 @@ int moduloMultiplication(int, int, int);
 int moduloPow(int, int,int);
 void brute_force(vector<vector<int>>*, vector<vector<int>>*, vector<int>*, vector<double>*);
 void curves_brute_force(vector<vector<double*>>*, vector<vector<double*>>*, vector<double>*, vector<double>*);
-double* arg_min(double**, vector<int>*, double, int);
+double* arg_min(double**, vector<double>*, double, int);
 double min(double, double, double);
