+++ lib/Traversals.h
@@ -2,6 +2,6 @@
 
 using namespace std;
 
-void shift_grid(vector<int>*, int, int);
-void hash_curve(vector<double*>*, vector<double*>*, vector<int>*, double, int);
+void shift_grid(vector<double>*, int, int);
+void hash_curve(vector<double*>*, vector<double*>*, vector<double>*, double, int);
 void Relevant_Traversals(vector<int*>*, int, int, int, int);
